requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openrouteservice==2.3.3
requests==2.32.3
